requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
APScheduler==3.11.2
lxml==4.9.3
selenium>=4.16.0
chromedriver-autoinstaller>=0.6.4
